+++ cppForSwig/DataObject.cpp
@@ -6,7 +6,23 @@
 //                                                                            //
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "BDM_seder.h"
+#include "DataObject.h"
+
+///////////////////////////////////////////////////////////////////////////////
+ostream& operator << (ostream& os, const ErrorType& et)
+{
+   os << et.err_;
+
+   return os;
+}
+
+///////////////////////////////////////////////////////////////////////////////
+istream& operator >> (istream& is, ErrorType& et)
+{
+   is >> et.err_;
+
+   return is;
+}
 
 ///////////////////////////////////////////////////////////////////////////////
 const vector<shared_ptr<DataMeta>> DataMeta::iTypeIDs_ = {
